Move `Dispatcher` to its own module
